+++ app/@modal/(...)auth/login/page.tsx
@@ -1,4 +1,4 @@
-import LoginModal from '@/features/auth/login/ui/login-form';
+import LoginModal from '@/features/auth/login/ui/login-modal';
 
 const Page = () => {
   return <LoginModal />;
